refactor: 🎨 link plugin moved to seperate file
